requirements.txt
neo4j
qdrant-client
sentence-transformers
python-dotenv
pandas
requests
